Update README.md
To use this script, add the class ".article-quiz" to the form element. Once your radio and/or checkbox options are setup, prefix each value with cf- or wf-. The dash is required, however you have the option to change the default name scheme. The status messages can be changed by updating the pseduo-selector content property.